+++ jest.config.cjs
@@ -1,4 +1,6 @@
+// eslint-disable-next-line @typescript-eslint/no-require-imports
 const nextJest = require("next/jest");
+
 const createJestConfig = nextJest({
   dir: "./",
 });